Move types and constants into separate files
