+++ sdks/node-sdk/src/types.ts
@@ -1,0 +1,65 @@
+import type { ContentCodec } from "@xmtp/content-type-primitives";
+import type { LogLevel } from "@xmtp/node-bindings";
+import type { ApiUrls } from "@/constants";
+
+/**
+ * XMTP environment
+ */
+export type XmtpEnv = keyof typeof ApiUrls;
+
+/**
+ * Network options
+ */
+export type NetworkOptions = {
+  /**
+   * Specify which XMTP environment to connect to. (default: `dev`)
+   */
+  env?: XmtpEnv;
+  /**
+   * apiUrl can be used to override the `env` flag and connect to a
+   * specific endpoint
+   */
+  apiUrl?: string;
+  /**
+   * historySyncUrl can be used to override the `env` flag and connect to a
+   * specific endpoint for syncing history
+   */
+  historySyncUrl?: string;
+};
+
+/**
+ * Storage options
+ */
+export type StorageOptions = {
+  /**
+   * Path to the local DB
+   */
+  dbPath?: string;
+};
+
+export type ContentOptions = {
+  /**
+   * Allow configuring codecs for additional content types
+   */
+  codecs?: ContentCodec[];
+};
+
+export type OtherOptions = {
+  /**
+   * Enable structured JSON logging
+   */
+  structuredLogging?: boolean;
+  /**
+   * Logging level
+   */
+  loggingLevel?: LogLevel;
+  /**
+   * Disable automatic registration when creating a client
+   */
+  disableAutoRegister?: boolean;
+};
+
+export type ClientOptions = NetworkOptions &
+  StorageOptions &
+  ContentOptions &
+  OtherOptions;
